Revert "feat: add sound picker UI and AlarmSound enum model"
This reverts commit 6650d5aaa9234c15db6f9a71104f3e0c2af6b65e.
