+++ LabelTimer/LabelTimer/Views/SettingsView.swift
@@ -32,7 +32,7 @@ struct SettingsView: View {
                     }
 
                     NavigationLink("기본 사운드") {
-                        SoundPickerView()
+                        Text("기본 사운드 선택 화면") // TODO: 연결 예정
                     }
 
                     Toggle("다크 모드", isOn: $isDarkMode)
